Fix for unsigned int and int

+++ src/util.c
@@ -2,7 +2,7 @@
 #include "libsbmlsim/libsbmlsim.h"
 
 int get_end_cycle(double sim_time, double dt) {
-  int r = sim_time / dt;
+  int r = (int)(sim_time / dt);
   return r;
 }
